Agents/requirements.txt
beautifulsoup4
strands-agents
strands-agents-tools
uv
uvicorn
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
aws-opentelemetry-distro